Exemplary create sequence
The sequence is referred to, but isn't created as far as I see it. I
rewrote the class definition for one table following `Associating a
Sequence as the Server Side Default` from the SQLAlchemy docs. The
sequence will be created, but I'm not sure the sequence is setup as
intended.